Update issue 246 SALVIA:   Fixed FreeImage/LibRawLite compiling bugs of mingw.   Fixed FreeImage/LibTIFF4 compiling bugs of mingw.   Fixed FreeImage/OpenEXR compiling bugs of mingw.   The value of enum "arch" and "systems" are string.   Refactor blibs/project.py for check and support mingw.   Refactor build script to support error report.   Support make bjam when mingw is used.   Code of configure and make project was shared.   Add toolchain error detection.   Improve unsupported toolchain and system detection.   Rename some configuration item of cmake.   Fixed bugs of add_boost_lib.   Fixed link error of LLVM.   Fixed parameter error of mingw.   Fixed gcc detection error when GCC is not in env PATH.   Rename eflib/include/ext_intrinsics.h to intrin.h   Add _bit_scan_forward/reverse to intrin   Didn't support sse cast any longer(SSE4 have).   Fixed compiling bugs of mingw of eflib.   Add cross-platform aligned_malloc and aligned_free.   Add cross-platform cpuid support.   Fixed compilation bugs of mingw of salviar/salviax/salviau/sasl.

+++ salviar/include/render_state.h
@@ -56,7 +56,7 @@ struct render_state
 	int32_t						base_vertex;
 	uint32_t					start_index;
 	uint32_t					prim_count;
-	
+
 	stream_state			    str_state;
 	input_layout_ptr			layout;
 
@@ -85,7 +85,7 @@ struct render_state
 	std::vector<surface_ptr>	color_targets;
 	surface_ptr					depth_stencil_target;
 
-    async_object_ptr			asyncs[async_object_ids::count];
+    async_object_ptr			asyncs[static_cast<int32_t>(async_object_ids::count)];
     async_object_ptr            current_async;
 
     viewport                    target_vp;
